@@ -31,9 +31,13 @@ import (
 // AuthEngineMountSpec defines the desired state of AuthEngineMount
 type AuthEngineMountSpec struct {
 
-	// Authentication is the kube aoth configuraiton to be used to execute this request
+	// Connection represents the information needed to connect to Vault. This operator uses the standard Vault environment variables to connect to Vault. If you need to override those settings and for example connect to a different Vault instance, you can do with this section of the CR.
+	// +kubebuilder:validation:Optional
+	Connection *vaultutils.VaultConnection `json:"connection,omitempty"`
+
+	// Authentication is the kube auth configuration to be used to execute this request
 	// +kubebuilder:validation:Required
-	Authentication KubeAuthConfiguration `json:"authentication,omitempty"`
+	Authentication vaultutils.KubeAuthConfiguration `json:"authentication,omitempty"`
 
 	AuthMount `json:",inline"`
 
@@ -41,7 +45,7 @@ type AuthEngineMountSpec struct {
 	// The final path will be {[spec.authentication.namespace]}/auth/{spec.path}/{metadata.name}.
 	// The authentication role must have the following capabilities = [ "create", "read", "update", "delete"] on that path /sys/auth/{[spec.authentication.namespace]}/{spec.path}/{metadata.name}.
 	// +kubebuilder:validation:Required
-	Path Path `json:"path,omitempty"`
+	Path vaultutils.Path `json:"path,omitempty"`
 }
 
 type AuthMount struct {
@@ -150,8 +154,16 @@ func (m *AuthMount) toMap() map[string]interface{} {
 	}
 }
 
+func (d *AuthEngineMount) GetVaultConnection() *vaultutils.VaultConnection {
+	return d.Spec.Connection
+}
+
+func (d *AuthEngineMount) GetKubeAuthConfiguration() *vaultutils.KubeAuthConfiguration {
+	return &d.Spec.Authentication
+}
+
 func (d *AuthEngineMount) GetPath() string {
-	return cleansePath(d.GetEngineListPath() + "/" + string(d.Spec.Path) + "/" + d.Name)
+	return vaultutils.CleansePath(d.GetEngineListPath() + "/" + string(d.Spec.Path) + "/" + d.Name)
 }
 
 func (d *AuthEngineMount) GetPayload() map[string]interface{} {
