@@ -39,15 +39,19 @@ type GitHubSecretEngineConfigSpec struct {
 	// INSERT ADDITIONAL SPEC FIELDS - desired state of cluster
 	// Important: Run "make" to regenerate code after modifying this file
 
-	// Authentication is the kube aoth configuraiton to be used to execute this request
+	// Connection represents the information needed to connect to Vault. This operator uses the standard Vault environment variables to connect to Vault. If you need to override those settings and for example connect to a different Vault instance, you can do with this section of the CR.
+	// +kubebuilder:validation:Optional
+	Connection *vaultutils.VaultConnection `json:"connection,omitempty"`
+
+	// Authentication is the kube auth configuration to be used to execute this request
 	// +kubebuilder:validation:Required
-	Authentication KubeAuthConfiguration `json:"authentication,omitempty"`
+	Authentication vaultutils.KubeAuthConfiguration `json:"authentication,omitempty"`
 
 	// Path at which to make the configuration.
 	// The final path will be {[spec.authentication.namespace]}/{spec.path}/config.
 	// The authentication role must have the following capabilities = [ "create", "read", "update", "delete"] on that path.
 	// +kubebuilder:validation:Required
-	Path Path `json:"path,omitempty"`
+	Path vaultutils.Path `json:"path,omitempty"`
 
 	GHConfig `json:",inline"`
 
@@ -61,14 +65,6 @@ type GHConfig struct {
 	// +kubebuilder:validation:Required
 	ApplicationID int64 `json:"applicationID,omitempty"`
 
-	// InstanceID the Installation ID of the GitHub App installation.
-	// +kubebuilder:validation:Optional
-	InstanceID int64 `json:"instanceID,omitempty"`
-
-	// OrganizationName the name of the organisation with the GitHub App installation.
-	// +kubebuilder:validation:Required
-	OrganizationName string `json:"organizationName,omitempty"`
-
 	// GitHubAPIBaseURL the base URL for API requests (defaults to the public GitHub API).
 	// +kubebuilder:validation:Optional
 	// +kubebuilder:default="https://api.github.com"
@@ -80,8 +76,6 @@ type GHConfig struct {
 func (i *GHConfig) toMap() map[string]interface{} {
 	payload := map[string]interface{}{}
 	payload["app_id"] = i.ApplicationID
-	payload["ins_id"] = i.InstanceID
-	payload["org_name"] = i.OrganizationName
 	payload["prv_key"] = i.retrievedSSHKey
 	payload["base_url"] = i.GitHubAPIBaseURL
 	return payload
@@ -90,7 +84,7 @@ func (i *GHConfig) toMap() map[string]interface{} {
 type SSHKeyConfig struct {
 	// VaultSecret retrieves the sshkey from a Vault secret. The sshkey will be retrieve at the key "key" (pun intented).
 	// +kubebuilder:validation:Optional
-	VaultSecret *VaultSecretReference `json:"vaultSecret,omitempty"`
+	VaultSecret *vaultutils.VaultSecretReference `json:"vaultSecret,omitempty"`
 
 	// Secret retrieves the ssh key from a Kubernetes secret. The secret must be of ssh type (https://kubernetes.io/docs/concepts/configuration/secret/#ssh-authentication-secrets).
 	// +kubebuilder:validation:Optional
@@ -98,6 +92,10 @@ type SSHKeyConfig struct {
 }
 
 var _ vaultutils.VaultObject = &GitHubSecretEngineConfig{}
+
+func (d *GitHubSecretEngineConfig) GetVaultConnection() *vaultutils.VaultConnection {
+	return d.Spec.Connection
+}
 
 func (d *GitHubSecretEngineConfig) GetPath() string {
 	return string(d.Spec.Path) + "/" + "config"
@@ -217,7 +215,11 @@ func (r *GitHubSecretEngineConfig) validateEitherFromVaultSecretOrFromSecret() e
 		count++
 	}
 	if count != 1 {
-		return errors.New("Only one of spec.sSHKeyReference.vaultSecret or spec.sSHKeyReference.secret can be specified.")
+		return errors.New("only one of spec.sSHKeyReference.vaultSecret or spec.sSHKeyReference.secret can be specified")
 	}
 	return nil
+}
+
+func (d *GitHubSecretEngineConfig) GetKubeAuthConfiguration() *vaultutils.KubeAuthConfiguration {
+	return &d.Spec.Authentication
 }
