@@ -32,15 +32,20 @@ import (
 
 // KubernetesAuthEngineConfigSpec defines the desired state of KubernetesAuthEngineConfig
 type KubernetesAuthEngineConfigSpec struct {
-	// Authentication is the kube aoth configuraiton to be used to execute this request
+
+	// Connection represents the information needed to connect to Vault. This operator uses the standard Vault environment variables to connect to Vault. If you need to override those settings and for example connect to a different Vault instance, you can do with this section of the CR.
+	// +kubebuilder:validation:Optional
+	Connection *vaultutils.VaultConnection `json:"connection,omitempty"`
+
+	// Authentication is the kube auth configuration to be used to execute this request
 	// +kubebuilder:validation:Required
-	Authentication KubeAuthConfiguration `json:"authentication,omitempty"`
+	Authentication vaultutils.KubeAuthConfiguration `json:"authentication,omitempty"`
 
 	// Path at which to make the configuration.
 	// The final path will be {[spec.authentication.namespace]}/auth/{spec.path}/config/{metadata.name}.
 	// The authentication role must have the following capabilities = [ "create", "read", "update", "delete"] on that path.
 	// +kubebuilder:validation:Required
-	Path Path `json:"path,omitempty"`
+	Path vaultutils.Path `json:"path,omitempty"`
 
 	KAECConfig `json:",inline"`
 
@@ -50,8 +55,12 @@ type KubernetesAuthEngineConfigSpec struct {
 	TokenReviewerServiceAccount *corev1.LocalObjectReference `json:"tokenReviewerServiceAccount,omitempty"`
 }
 
+func (d *KubernetesAuthEngineConfig) GetVaultConnection() *vaultutils.VaultConnection {
+	return d.Spec.Connection
+}
+
 func (d *KubernetesAuthEngineConfig) GetPath() string {
-	return cleansePath("auth/" + string(d.Spec.Path) + "/" + d.Name + "/config")
+	return vaultutils.CleansePath("auth/" + string(d.Spec.Path) + "/" + d.Name + "/config")
 }
 
 func (d *KubernetesAuthEngineConfig) GetPayload() map[string]interface{} {
@@ -80,7 +89,8 @@ func (d *KubernetesAuthEngineConfig) PrepareInternalValues(context context.Conte
 }
 
 func (kc *KubernetesAuthEngineConfig) getJWTToken(context context.Context) (string, error) {
-	return getJWTToken(context, kc.Spec.TokenReviewerServiceAccount.Name, kc.Namespace)
+	expiration := int64(60 * 60 * 24 * 365)
+	return vaultutils.GetJWTTokenWithDuration(context, kc.Spec.TokenReviewerServiceAccount.Name, kc.Namespace, expiration)
 }
 
 func (r *KubernetesAuthEngineConfig) IsValid() (bool, error) {
@@ -173,4 +183,8 @@ func (i *KAECConfig) toMap() map[string]interface{} {
 	payload["disable_iss_validation"] = i.DisableISSValidation
 	payload["disable_local_ca_jwt"] = i.DisableLocalCAJWT
 	return payload
+}
+
+func (d *KubernetesAuthEngineConfig) GetKubeAuthConfiguration() *vaultutils.KubeAuthConfiguration {
+	return &d.Spec.Authentication
 }
