@@ -34,40 +34,35 @@ import (
 // KubernetesAuthEngineRoleSpec defines the desired state of KubernetesAuthEngineRole
 type KubernetesAuthEngineRoleSpec struct {
 
-	// Authentication is the kube aoth configuraiton to be used to execute this request
+	// Connection represents the information needed to connect to Vault. This operator uses the standard Vault environment variables to connect to Vault. If you need to override those settings and for example connect to a different Vault instance, you can do with this section of the CR.
+	// +kubebuilder:validation:Optional
+	Connection *vaultutils.VaultConnection `json:"connection,omitempty"`
+
+	// Authentication is the kube auth configuration to be used to execute this request
 	// +kubebuilder:validation:Required
-	Authentication KubeAuthConfiguration `json:"authentication,omitempty"`
+	Authentication vaultutils.KubeAuthConfiguration `json:"authentication,omitempty"`
 
 	// Path at which to make the configuration.
 	// The final path will be {[spec.authentication.namespace]}/auth/{spec.path}/role/{metadata.name}.
 	// The authentication role must have the following capabilities = [ "create", "read", "update", "delete"] on that path.
 	// +kubebuilder:validation:Required
-	Path Path `json:"path,omitempty"`
+	Path vaultutils.Path `json:"path,omitempty"`
 
 	VRole `json:",inline"`
 
 	// TargetNamespaces specifies how to retrieve the namespaces bound to this Vault role.
 	// +kubebuilder:validation:Required
-	TargetNamespaces TargetNamespaceConfig `json:"targetNamespaces,omitempty"`
-}
-
-type TargetNamespaceConfig struct {
-	// TargetNamespaceSelector is a selector of namespaces from which service accounts will receove this role. Either TargetNamespaceSelector or TargetNamespaces can be specified
-	// +kubebuilder:validation:Optional
-	TargetNamespaceSelector *metav1.LabelSelector `json:"targetNamespaceSelector,omitempty"`
-
-	// TargetNamespaces is a list of namespace from which service accounts will receive this role. Either TargetNamespaceSelector or TargetNamespaces can be specified.
-	// +kubebuilder:validation:Optional
-	// +kubebuilder:validation:MinItems=1
-	// kubebuilder:validation:UniqueItems=true
-	// +listType=set
-	TargetNamespaces []string `json:"targetNamespaces,omitempty"`
+	TargetNamespaces vaultutils.TargetNamespaceConfig `json:"targetNamespaces,omitempty"`
 }
 
 var _ vaultutils.VaultObject = &KubernetesAuthEngineRole{}
 
+func (d *KubernetesAuthEngineRole) GetVaultConnection() *vaultutils.VaultConnection {
+	return d.Spec.Connection
+}
+
 func (d *KubernetesAuthEngineRole) GetPath() string {
-	return cleansePath("auth/" + string(d.Spec.Path) + "/role/" + d.Name)
+	return vaultutils.CleansePath("auth/" + string(d.Spec.Path) + "/role/" + d.Name)
 }
 func (d *KubernetesAuthEngineRole) GetPayload() map[string]interface{} {
 	return d.Spec.VRole.toMap()
@@ -89,7 +84,12 @@ func (d *KubernetesAuthEngineRole) PrepareInternalValues(context context.Context
 			log.Error(err, "unable to retrieve selected namespaces", "instance", object)
 			return err
 		}
-		d.SetInternalNamespaces(namespaces)
+		if len(namespaces) < 1 {
+			// workaround for when there are no namespace, as we don't want this call to error out
+			d.SetInternalNamespaces([]string{"__no_namespace__"})
+		} else {
+			d.SetInternalNamespaces(namespaces)
+		}
 	} else {
 		d.SetInternalNamespaces(d.Spec.TargetNamespaces.TargetNamespaces)
 	}
@@ -132,20 +132,26 @@ type VRole struct {
 	// +kubebuilder:default={"default"}
 	TargetServiceAccounts []string `json:"targetServiceAccounts"`
 
-	// Policies is a list of policy names to be bound to this role.
-	// +kubebuilder:validation:MinItems=1
-	// kubebuilder:validation:UniqueItems=true
-	// +kubebuilder:validation:Required
-	Policies []string `json:"policies"`
-
 	// Audience Audience claim to verify in the JWT.
 	// +kubebuilder:validation:Optional
-	Audience string `json:"audience,omitempty"`
+	Audience *string `json:"audience,omitempty"`
+
+	// AliasNameSource Configures how identity aliases are generated. Valid choices are: serviceaccount_uid, serviceaccount_name When serviceaccount_uid is specified, the machine generated UID from the service account will be used as the identity alias name. When serviceaccount_name is specified, the service account's namespace and name will be used as the identity alias name e.g vault/vault-auth. While it is strongly advised that you use serviceaccount_uid, you may also use serviceaccount_name in cases where you want to set the alias ahead of time, and the risks are mitigated or otherwise acceptable given your use case. It is very important to limit who is able to delete/create service accounts within a given cluster. See the Create an Entity Alias document which further expands on the potential security implications mentioned above.
+	// +kubebuilder:validation:Optional
+	// +kubebuilder:validation:Enum:={"serviceaccount_uid", "serviceaccount_name"}
+	// +kubebuilder:default="serviceaccount_uid"
+	AliasNameSource string `json:"aliasNameSource,omitempty"`
 
 	// TokenTTL The incremental lifetime for generated tokens. This current value of this will be referenced at renewal time.
 	// +kubebuilder:validation:Optional
 	// +kubebuilder:default:=0
 	TokenTTL int `json:"tokenTTL,omitempty"`
+
+	// Policies is a list of policy names to be bound to this role.
+	// +kubebuilder:validation:MinItems=1
+	// kubebuilder:validation:UniqueItems=true
+	// +kubebuilder:validation:Required
+	Policies []string `json:"policies"`
 
 	// TokenMaxTTL The maximum lifetime for generated tokens. This current value of this will be referenced at renewal time.
 	// +kubebuilder:validation:Optional
@@ -181,6 +187,7 @@ type VRole struct {
 	// TokenType The type of token that should be generated. Can be service, batch, or default to use the mount's tuned default (which unless changed will be service tokens). For token store roles, there are two additional possibilities: default-service and default-batch which specify the type to return unless the client requests a different type at generation time.
 	// +kubebuilder:validation:Optional
 	// +kubebuilder:validation:Enum:={"service","batch","default","default-service","default-batch"}
+	// +kubebuilder:default="default"
 	TokenType string `json:"tokenType,omitempty"`
 
 	// this field is for internal use and will not be serialized
@@ -238,7 +245,10 @@ func (i *VRole) toMap() map[string]interface{} {
 	payload := map[string]interface{}{}
 	payload["bound_service_account_names"] = i.TargetServiceAccounts
 	payload["bound_service_account_namespaces"] = i.namespaces
-	payload["audience"] = i.Audience
+	payload["alias_name_source"] = i.AliasNameSource
+	if i.Audience != nil {
+		payload["audience"] = i.Audience
+	}
 	payload["token_ttl"] = i.TokenTTL
 	payload["token_max_ttl"] = i.TokenMaxTTL
 	payload["token_policies"] = i.Policies
@@ -246,7 +256,7 @@ func (i *VRole) toMap() map[string]interface{} {
 	payload["token_explicit_max_ttl"] = i.TokenExplicitMaxTTL
 	payload["token_no_default_policy"] = i.TokenNoDefaultPolicy
 	payload["token_num_uses"] = i.TokenNumUses
-	payload["tokenPeriod"] = i.TokenPeriod
+	payload["token_period"] = i.TokenPeriod
 	payload["token_type"] = i.TokenType
 	return payload
 }
@@ -264,7 +274,11 @@ func (r *KubernetesAuthEngineRole) validateEitherTargetNamespaceSelectorOrTarget
 		count++
 	}
 	if count != 1 {
-		return errors.New("Only one of TargetNamespaceSelector or TargetNamespaces can be specified.")
+		return errors.New("only one of TargetNamespaceSelector or TargetNamespaces can be specified")
 	}
 	return nil
+}
+
+func (d *KubernetesAuthEngineRole) GetKubeAuthConfiguration() *vaultutils.KubeAuthConfiguration {
+	return &d.Spec.Authentication
 }
