@@ -28,15 +28,19 @@ import (
 // LDAPAuthEngineGroupSpec defines the desired state of LDAPAuthEngineGroup
 type LDAPAuthEngineGroupSpec struct {
 
+	// Connection represents the information needed to connect to Vault. This operator uses the standard Vault environment variables to connect to Vault. If you need to override those settings and for example connect to a different Vault instance, you can do with this section of the CR.
+	// +kubebuilder:validation:Optional
+	Connection *vaultutils.VaultConnection `json:"connection,omitempty"`
+
 	// Authentication is the kube auth configuraiton to be used to execute this request
 	// +kubebuilder:validation:Required
-	Authentication KubeAuthConfiguration `json:"authentication,omitempty"`
+	Authentication vaultutils.KubeAuthConfiguration `json:"authentication,omitempty"`
 
 	// Path at which to make the configuration.
 	// The final path will be {[spec.authentication.namespace]}/auth/{spec.path}/groups/{metadata.name}.
 	// The authentication role must have the following capabilities = [ "create", "read", "update", "delete"] on that path.
 	// +kubebuilder:validation:Required
-	Path Path `json:"path,omitempty"`
+	Path vaultutils.Path `json:"path,omitempty"`
 
 	// The name of the LDAP group
 	// +kubebuilder:validation:Required
@@ -50,8 +54,12 @@ type LDAPAuthEngineGroupSpec struct {
 
 var _ vaultutils.VaultObject = &LDAPAuthEngineGroup{}
 
+func (d *LDAPAuthEngineGroup) GetVaultConnection() *vaultutils.VaultConnection {
+	return d.Spec.Connection
+}
+
 func (d *LDAPAuthEngineGroup) GetPath() string {
-	return cleansePath("auth/" + string(d.Spec.Path) + "/groups/" + string(d.Spec.Name))
+	return vaultutils.CleansePath("auth/" + string(d.Spec.Path) + "/groups/" + string(d.Spec.Name))
 }
 
 func (d *LDAPAuthEngineGroup) GetPayload() map[string]interface{} {
@@ -122,4 +130,8 @@ func (i *LDAPAuthEngineGroup) toMap() map[string]interface{} {
 	payload["policies"] = i.Spec.Policies
 
 	return payload
+}
+
+func (d *LDAPAuthEngineGroup) GetKubeAuthConfiguration() *vaultutils.KubeAuthConfiguration {
+	return &d.Spec.Authentication
 }
