@@ -30,7 +30,14 @@ import (
 
 var _ vaultutils.VaultObject = &Policy{}
 
+func (d *Policy) GetVaultConnection() *vaultutils.VaultConnection {
+	return d.Spec.Connection
+}
+
 func (d *Policy) GetPath() string {
+	if d.Spec.Type != "" {
+		return "sys/policies/" + d.Spec.Type + "/" + d.Name
+	}
 	return "sys/policy/" + d.Name
 }
 func (d *Policy) GetPayload() map[string]interface{} {
@@ -39,7 +46,13 @@ func (d *Policy) GetPayload() map[string]interface{} {
 	}
 }
 func (d *Policy) IsEquivalentToDesiredState(payload map[string]interface{}) bool {
-	return reflect.DeepEqual(d.GetPayload(), payload)
+	desiredState := d.GetPayload()
+	desiredState["name"] = d.Name
+	if d.Spec.Type == "" {
+		desiredState["rules"] = desiredState["policy"]
+		delete(desiredState, "policy")
+	}
+	return reflect.DeepEqual(desiredState, payload)
 }
 
 func (d *Policy) IsInitialized() bool {
@@ -59,13 +72,22 @@ type PolicySpec struct {
 	// INSERT ADDITIONAL SPEC FIELDS - desired state of cluster
 	// Important: Run "make" to regenerate code after modifying this file
 
+	// Connection represents the information needed to connect to Vault. This operator uses the standard Vault environment variables to connect to Vault. If you need to override those settings and for example connect to a different Vault instance, you can do with this section of the CR.
+	// +kubebuilder:validation:Optional
+	Connection *vaultutils.VaultConnection `json:"connection,omitempty"`
+
 	// Policy is a Vault policy expressed in HCL language.
 	// +kubebuilder:validation:Required
 	Policy string `json:"policy,omitempty"`
 
-	// Authentication is the kube aoth configuraiton to be used to execute this request
+	// Type represents the policy type, currently the only supported policy type is "acl", but in the future rgp and egp  might be supported. If not specified a policy will be created at /sys/policies/<name>, if specified (the recommended approach) a policy will be created at /sys/policies/acl/<name>
+	// +kubebuilder:validation:Optional
+	// +kubebuilder:validation:Enum={"acl"}
+	Type string `json:"type,omitempty"`
+
+	// Authentication is the kube auth configuration to be used to execute this request
 	// +kubebuilder:validation:Required
-	Authentication KubeAuthConfiguration `json:"authentication,omitempty"`
+	Authentication vaultutils.KubeAuthConfiguration `json:"authentication,omitempty"`
 }
 
 // PolicyStatus defines the observed state of Policy
@@ -109,4 +131,8 @@ type PolicyList struct {
 
 func init() {
 	SchemeBuilder.Register(&Policy{}, &PolicyList{})
+}
+
+func (d *Policy) GetKubeAuthConfiguration() *vaultutils.KubeAuthConfiguration {
+	return &d.Spec.Authentication
 }
