@@ -33,24 +33,32 @@ import (
 // QuaySecretEngineConfigSpec defines the desired state of QuaySecretEngineConfig
 type QuaySecretEngineConfigSpec struct {
 
+	// Connection represents the information needed to connect to Vault. This operator uses the standard Vault environment variables to connect to Vault. If you need to override those settings and for example connect to a different Vault instance, you can do with this section of the CR.
+	// +kubebuilder:validation:Optional
+	Connection *vaultutils.VaultConnection `json:"connection,omitempty"`
+
 	// Authentication is the kube auth configuration to be used to execute this request
 	// +kubebuilder:validation:Required
-	Authentication KubeAuthConfiguration `json:"authentication,omitempty"`
+	Authentication vaultutils.KubeAuthConfiguration `json:"authentication,omitempty"`
 
 	// Path at which to make the configuration.
 	// The final path will be {[spec.authentication.namespace]}/{spec.path}/config.
 	// The authentication role must have the following capabilities = [ "create", "read", "update", "delete"] on that path.
 	// +kubebuilder:validation:Required
-	Path Path `json:"path,omitempty"`
+	Path vaultutils.Path `json:"path,omitempty"`
 
 	QuayConfig `json:",inline"`
 
 	// RootCredentials specifies how to retrieve the credentials for this Quay connection.
 	// +kubebuilder:validation:Required
-	RootCredentials RootCredentialConfig `json:"rootCredentials,omitempty"`
+	RootCredentials vaultutils.RootCredentialConfig `json:"rootCredentials,omitempty"`
 }
 
 var _ vaultutils.VaultObject = &QuaySecretEngineConfig{}
+
+func (d *QuaySecretEngineConfig) GetVaultConnection() *vaultutils.VaultConnection {
+	return d.Spec.Connection
+}
 
 func (q *QuaySecretEngineConfig) GetPath() string {
 	return string(q.Spec.Path) + "/" + "config"
@@ -210,5 +218,9 @@ func init() {
 }
 
 func (r *QuaySecretEngineConfig) isValid() error {
-	return r.Spec.RootCredentials.validateEitherFromVaultSecretOrFromSecretOrFromRandomSecret()
+	return r.Spec.RootCredentials.ValidateEitherFromVaultSecretOrFromSecretOrFromRandomSecret()
+}
+
+func (d *QuaySecretEngineConfig) GetKubeAuthConfiguration() *vaultutils.KubeAuthConfiguration {
+	return &d.Spec.Authentication
 }
