@@ -49,20 +49,28 @@ const (
 // QuaySecretEngineRoleSpec defines the desired state of QuaySecretEngineRole
 type QuaySecretEngineRoleSpec struct {
 
+	// Connection represents the information needed to connect to Vault. This operator uses the standard Vault environment variables to connect to Vault. If you need to override those settings and for example connect to a different Vault instance, you can do with this section of the CR.
+	// +kubebuilder:validation:Optional
+	Connection *vaultutils.VaultConnection `json:"connection,omitempty"`
+
 	// Authentication is the kube auth configuration to be used to execute this request
 	// +kubebuilder:validation:Required
-	Authentication KubeAuthConfiguration `json:"authentication,omitempty"`
+	Authentication vaultutils.KubeAuthConfiguration `json:"authentication,omitempty"`
 
 	// Path at which to make the configuration.
 	// The final path will be {[spec.authentication.namespace]}/{spec.path}/roles/{metadata.name}.
 	// The authentication role must have the following capabilities = [ "create", "read", "update", "delete"] on that path.
 	// +kubebuilder:validation:Required
-	Path Path `json:"path,omitempty"`
+	Path vaultutils.Path `json:"path,omitempty"`
 
 	QuayRole `json:",inline"`
 }
 
 var _ vaultutils.VaultObject = &QuaySecretEngineRole{}
+
+func (d *QuaySecretEngineRole) GetVaultConnection() *vaultutils.VaultConnection {
+	return d.Spec.Connection
+}
 
 func (q *QuaySecretEngineRole) GetPath() string {
 	return string(q.Spec.Path) + "/" + "roles" + "/" + q.Name
@@ -198,4 +206,8 @@ type QuaySecretEngineRoleList struct {
 
 func init() {
 	SchemeBuilder.Register(&QuaySecretEngineRole{}, &QuaySecretEngineRoleList{})
+}
+
+func (d *QuaySecretEngineRole) GetKubeAuthConfiguration() *vaultutils.KubeAuthConfiguration {
+	return &d.Spec.Authentication
 }
