@@ -30,15 +30,20 @@ import (
 
 // RabbitMQSecretEngineRoleSpec defines the desired state of RabbitMQSecretEngineRole
 type RabbitMQSecretEngineRoleSpec struct {
+
+	// Connection represents the information needed to connect to Vault. This operator uses the standard Vault environment variables to connect to Vault. If you need to override those settings and for example connect to a different Vault instance, you can do with this section of the CR.
+	// +kubebuilder:validation:Optional
+	Connection *vaultutils.VaultConnection `json:"connection,omitempty"`
+
 	// Authentication is the k8s auth configuration to be used to execute this request
 	// +kubebuilder:validation:Required
-	Authentication KubeAuthConfiguration `json:"authentication"`
+	Authentication vaultutils.KubeAuthConfiguration `json:"authentication"`
 
 	// Path at which to make the configuration.
 	// The final path will be {[spec.authentication.namespace]}/{spec.path}/config/{metadata.name}.
 	// The authentication role must have the following capabilities = [ "create", "read", "update", "delete"] on that path.
 	// +kubebuilder:validation:Required
-	Path Path `json:"path"`
+	Path vaultutils.Path `json:"path"`
 
 	// +kubebuilder:validation:Required
 	RMQSERole `json:",inline"`
@@ -135,6 +140,10 @@ type RabbitMQSecretEngineRoleList struct {
 
 var _ apis.ConditionsAware = &RabbitMQSecretEngineConfig{}
 
+func (d *RabbitMQSecretEngineRole) GetVaultConnection() *vaultutils.VaultConnection {
+	return d.Spec.Connection
+}
+
 func (m *RabbitMQSecretEngineRole) GetConditions() []metav1.Condition {
 	return m.Status.Conditions
 }
@@ -212,4 +221,8 @@ func (fields *RMQSERole) rabbitMQToMap() map[string]interface{} {
 	payload["vhosts"] = convertVhostsToJson(fields.Vhosts)
 	payload["vhost_topics"] = convertTopicsToJson(fields.VhostTopics)
 	return payload
+}
+
+func (d *RabbitMQSecretEngineRole) GetKubeAuthConfiguration() *vaultutils.KubeAuthConfiguration {
+	return &d.Spec.Authentication
 }
