@@ -14,24 +14,28 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package v1alpha1
+package utils
 
 import (
 	"context"
 	"errors"
 	"strings"
-	"time"
 
 	vault "github.com/hashicorp/vault/api"
+	authv1 "k8s.io/api/authentication/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/client-go/kubernetes"
+	"k8s.io/client-go/rest"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/log"
 )
 
+// +kubebuilder:object:generate=true
 // +kubebuilder:validation:Pattern:=`^(?:/?[\w;:@&=\$-\.\+]*)+/?`
 type Path string
 
+// +kubebuilder:object:generate=true
 type KubeAuthConfiguration struct {
 	// ServiceAccount is the service account used for the kube auth authentication
 	// +kubebuilder:validation:Required
@@ -52,6 +56,89 @@ type KubeAuthConfiguration struct {
 	Namespace string `json:"namespace,omitempty"`
 }
 
+// +kubebuilder:object:generate=true
+type VaultConnection struct {
+	// +kubebuilder:validation:Optional
+	TLSConfig *TLSConfig `json:"tLSConfig,omitempty"`
+
+	// Address Address of the Vault server expressed as a URL and port, for example: https://127.0.0.1:8200/
+	// +kubebuilder:validation:Required
+	Address string `json:"address,omitempty"`
+
+	// Timeout Timeout variable. The default value is 60s.
+	// +kubebuilder:validation:Optional
+	TimeOut *metav1.Duration `json:"timeOut,omitempty"`
+
+	// MaxRetries Maximum number of retries when certain error codes are encountered. The default is 2, for three total attempts. Set this to 0 or less to disable retrying. Error codes that are retried are 412 (client consistency requirement not satisfied) and all 5xx except for 501 (not implemented).
+	// +kubebuilder:validation:Optional
+	MaxRetries *int `json:"maxRetries,omitempty"`
+}
+
+// +kubebuilder:object:generate=true
+type TLSConfig struct {
+	// Cacert Path to a PEM-encoded CA certificate file on the local disk. This file is used to verify the Vault server's SSL certificate. This environment variable takes precedence over a cert passed via the secret.
+	// +kubebuilder:validation:Optional
+	Cacert *string `json:"cacert,omitempty"`
+
+	// TLSSecret namespace-local secret containing the tls material for the connection. the expected keys for the secret are: ca bundle -> "ca.crt", certificate -> "tls.crt", key -> "tls.key"
+	// +kubebuilder:validation:Optional
+	TLSSecret *corev1.LocalObjectReference `json:"tlsSecret,omitempty"`
+
+	// SkipVerify Do not verify Vault's presented certificate before communicating with it. Setting this variable is not recommended and voids Vault's security model.
+	// +kubebuilder:validation:Optional
+	SkipVerify bool `json:"skipVerify,omitempty"`
+
+	// TLSServerName Name to use as the SNI host when connecting via TLS.
+	// +kubebuilder:validation:Optional
+	TLSServerName *string `json:"tlsServerName,omitempty"`
+}
+
+func (vc *VaultConnection) getConnectionConfig(context context.Context, kubeNamespace string) (*vault.Config, error) {
+	log := log.FromContext(context)
+	restConfig := context.Value("restConfig").(*rest.Config)
+	clientset, err := kubernetes.NewForConfig(restConfig)
+	if err != nil {
+		log.Error(err, "unable to create kubernetes clientset")
+		return nil, err
+	}
+	config := vault.DefaultConfig()
+	config.Address = vc.Address
+	if vc.TimeOut != nil {
+		config.Timeout = vc.TimeOut.Duration
+	}
+	if vc.MaxRetries != nil {
+		config.MaxRetries = *vc.MaxRetries
+	}
+	if vc.TLSConfig != nil {
+		tlsConfig := vault.TLSConfig{}
+		if vc.TLSConfig.TLSSecret != nil {
+			tlsSecret, err := clientset.CoreV1().Secrets(kubeNamespace).Get(context, vc.TLSConfig.TLSSecret.Name, metav1.GetOptions{})
+			if err != nil {
+				log.Error(err, "unable to retrieve", "secret", vc.TLSConfig.TLSSecret.Name)
+				return nil, err
+			}
+			if ca, ok := tlsSecret.Data["ca.crt"]; ok {
+				tlsConfig.CACertBytes = ca
+			}
+			if key, ok := tlsSecret.Data["tls.key"]; ok {
+				tlsConfig.ClientKey = string(key)
+			}
+			if crt, ok := tlsSecret.Data["tls.crt"]; ok {
+				tlsConfig.ClientCert = string(crt)
+			}
+		}
+		if vc.TLSConfig.Cacert != nil {
+			tlsConfig.CACert = *vc.TLSConfig.Cacert
+		}
+		if vc.TLSConfig.TLSServerName != nil {
+			tlsConfig.TLSServerName = *vc.TLSConfig.TLSServerName
+		}
+		tlsConfig.Insecure = !vc.TLSConfig.SkipVerify
+		config.ConfigureTLS(&tlsConfig)
+	}
+	return config, nil
+}
+
 func (kc *KubeAuthConfiguration) GetNamespace() string {
 	return kc.Namespace
 }
@@ -59,7 +146,7 @@ func (kc *KubeAuthConfiguration) GetRole() string {
 	return kc.Role
 }
 func (kc *KubeAuthConfiguration) GetKubeAuthPath() string {
-	return cleansePath("auth/" + string(kc.Path) + "/login")
+	return CleansePath("auth/" + string(kc.Path) + "/login")
 }
 
 func (kc *KubeAuthConfiguration) GetServiceAccountName() string {
@@ -73,7 +160,7 @@ func (kc *KubeAuthConfiguration) GetVaultClient(context context.Context, kubeNam
 		log.Error(err, "unable to retrieve jwt token for", "namespace", kubeNamespace, "serviceaccount", kc.GetServiceAccountName())
 		return nil, err
 	}
-	vaultClient, err := kc.createVaultClient(context, jwt)
+	vaultClient, err := kc.createVaultClient(context, jwt, kubeNamespace)
 	if err != nil {
 		log.Error(err, "unable to create vault client")
 		return nil, err
@@ -81,81 +168,58 @@ func (kc *KubeAuthConfiguration) GetVaultClient(context context.Context, kubeNam
 	return vaultClient, nil
 }
 
-func getJWTToken(context context.Context, serviceAccountName string, kubeNamespace string) (string, error) {
+func GetJWTTokenWithDuration(context context.Context, serviceAccountName string, kubeNamespace string, duration int64) (string, error) {
 	log := log.FromContext(context)
-	kubeClient := context.Value("kubeClient").(client.Client)
-	secretList := &corev1.SecretList{}
-	err := kubeClient.List(context, secretList, &client.ListOptions{
-		Namespace: kubeNamespace,
-	})
-	if err != nil {
-		log.Error(err, "unable to retrieve secrets", "in namespace", kubeNamespace)
-		return "", err
-	}
-	for _, secret := range secretList.Items {
-		if saname, ok := secret.Annotations["kubernetes.io/service-account.name"]; ok {
-			if saname == serviceAccountName {
-				if jwt, ok := secret.Data["token"]; ok {
-					return string(jwt), nil
-				} else {
-					return "", errors.New("unable to find \"token\" key in secret" + kubeNamespace + "/" + secret.Name)
-				}
-			}
-		}
+
+	restConfig := context.Value("restConfig").(*rest.Config)
+
+	treq := &authv1.TokenRequest{
+		Spec: authv1.TokenRequestSpec{
+			ExpirationSeconds: &duration,
+		},
 	}
 
-	//if we get here, we were not able to find the secret with teh new method mandatory in kube 1.24 and supported in a few previous version.
-	//so we try the otgher approach by looking at the service account
+	clientset, err := kubernetes.NewForConfig(restConfig)
 
-	serviceAccount := &corev1.ServiceAccount{}
-	err = kubeClient.Get(context, client.ObjectKey{
-		Namespace: kubeNamespace,
-		Name:      serviceAccountName,
-	}, serviceAccount)
 	if err != nil {
-		log.Error(err, "unable to retrieve", "service account", client.ObjectKey{
-			Namespace: kubeNamespace,
-			Name:      serviceAccountName,
-		})
+		log.Error(err, "unable to create kubernetes clientset")
 		return "", err
 	}
 
-	var tokenSecretName string
-	for _, secretName := range serviceAccount.Secrets {
-		if strings.Contains(secretName.Name, "token") {
-			tokenSecretName = secretName.Name
-			break
-		}
-	}
-	if tokenSecretName == "" {
-		return "", errors.New("unable to find token secret name for service account" + kubeNamespace + "/" + serviceAccountName)
-	}
-	secret := &corev1.Secret{}
-	err = kubeClient.Get(context, client.ObjectKey{
-		Namespace: kubeNamespace,
-		Name:      tokenSecretName,
-	}, secret)
+	treq, err = clientset.CoreV1().ServiceAccounts(kubeNamespace).CreateToken(context, serviceAccountName, treq, metav1.CreateOptions{})
 	if err != nil {
-		log.Error(err, "unable to retrieve", "secret", client.ObjectKey{
-			Namespace: kubeNamespace,
-			Name:      tokenSecretName,
-		})
+		log.Error(err, "unable to create service account token request", "in namespace", kubeNamespace, "for service account", serviceAccountName)
 		return "", err
 	}
-	if jwt, ok := secret.Data["token"]; ok {
-		return string(jwt), nil
-	} else {
-		return "", errors.New("unable to find \"token\" key in secret" + kubeNamespace + "/" + tokenSecretName)
-	}
+
+	return treq.Status.Token, nil
+}
+
+func GetJWTToken(context context.Context, serviceAccountName string, kubeNamespace string) (string, error) {
+	expiration := int64(600)
+
+	return GetJWTTokenWithDuration(context, serviceAccountName, kubeNamespace, expiration)
 }
 
 func (kc *KubeAuthConfiguration) getJWTToken(context context.Context, kubeNamespace string) (string, error) {
-	return getJWTToken(context, kc.GetServiceAccountName(), kubeNamespace)
+	return GetJWTToken(context, kc.GetServiceAccountName(), kubeNamespace)
 }
 
-func (kc *KubeAuthConfiguration) createVaultClient(context context.Context, jwt string) (*vault.Client, error) {
+func (kc *KubeAuthConfiguration) createVaultClient(context context.Context, jwt string, namespace string) (*vault.Client, error) {
 	log := log.FromContext(context)
-	config := vault.DefaultConfig()
+	vaultConnection := context.Value("vaultConnection").(*VaultConnection)
+	var config *vault.Config
+	if vaultConnection != nil {
+		var err error
+		config, err = vaultConnection.getConnectionConfig(context, namespace)
+		if err != nil {
+			log.Error(err, "unable initialize vault connection configuration")
+			return nil, err
+		}
+	} else {
+		config = vault.DefaultConfig()
+	}
+
 	client, err := vault.NewClient(config)
 	if err != nil {
 		log.Error(err, "unable initialize vault client")
@@ -179,33 +243,25 @@ func (kc *KubeAuthConfiguration) createVaultClient(context context.Context, jwt 
 	return client, nil
 }
 
-func cleansePath(path string) string {
+func CleansePath(path string) string {
 	return strings.Trim(strings.ReplaceAll(path, "//", "/"), "/")
 }
 
-func toString(name interface{}) string {
+func ToString(name interface{}) string {
 	if name != nil {
 		return name.(string)
 	}
 	return ""
 }
 
-func parseOrDie(val string) metav1.Duration {
-	d, err := time.ParseDuration(val)
-	if err != nil {
-		panic(err)
-	}
-	return metav1.Duration{
-		Duration: d,
-	}
-}
-
+// +kubebuilder:object:generate=true
 type VaultSecretReference struct {
 	// Path is the path to the secret
 	// +kubebuilder:validation:Required
 	Path string `json:"path,omitempty"`
 }
 
+// +kubebuilder:object:generate=true
 type RootCredentialConfig struct {
 	// VaultSecret retrieves the credentials from a Vault secret. This will map the "username" and "password" keys of the secret to the username and password of this config. All other keys will be ignored. Only one of RootCredentialsFromVaultSecret or RootCredentialsFromSecret or RootCredentialsFromRandomSecret can be specified.
 	// username: Specifies the name of the user to use as the "root" user when connecting to the database. This "root" user is used to create/update/delete users managed by these plugins, so you will need to ensure that this user has permissions to manipulate users appropriate to the database. This is typically used in the connection_url field via the templating directive "{{"username"}}" or "{{"name"}}".
@@ -229,14 +285,30 @@ type RootCredentialConfig struct {
 
 	// PasswordKey key to be used when retrieving the password, required with VaultSecrets and Kubernetes secrets, ignored with RandomSecret
 	// +kubebuilder:validation:Optional
+	// +kubebuilder:default="password"
 	PasswordKey string `json:"passwordKey,omitempty"`
 
 	// UsernameKey key to be used when retrieving the username, optional with VaultSecrets and Kubernetes secrets, ignored with RandomSecret
 	// +kubebuilder:validation:Optional
+	// +kubebuilder:default="username"
 	UsernameKey string `json:"usernameKey,omitempty"`
 }
 
-func (credentials *RootCredentialConfig) validateEitherFromVaultSecretOrFromSecretOrFromRandomSecret() error {
+func (credentials *RootCredentialConfig) ValidateEitherFromVaultSecretOrFromSecret() error {
+	count := 0
+	if credentials.Secret != nil {
+		count++
+	}
+	if credentials.VaultSecret != nil {
+		count++
+	}
+	if count != 1 {
+		return errors.New("only one of spec.rootCredentials.vaultSecret or spec.rootCredentials.secret or spec.rootCredentials.randomSecret can be specified")
+	}
+	return nil
+}
+
+func (credentials *RootCredentialConfig) ValidateEitherFromVaultSecretOrFromSecretOrFromRandomSecret() error {
 	count := 0
 	if credentials.RandomSecret != nil {
 		count++
@@ -255,4 +327,18 @@ func (credentials *RootCredentialConfig) validateEitherFromVaultSecretOrFromSecr
 
 func GetFinalizer(instance client.Object) string {
 	return "controller-" + strings.ToLower(instance.GetObjectKind().GroupVersionKind().Kind)
+}
+
+// +kubebuilder:object:generate=true
+type TargetNamespaceConfig struct {
+	// TargetNamespaceSelector is a selector of namespaces from which service accounts will receove this role. Either TargetNamespaceSelector or TargetNamespaces can be specified
+	// +kubebuilder:validation:Optional
+	TargetNamespaceSelector *metav1.LabelSelector `json:"targetNamespaceSelector,omitempty"`
+
+	// TargetNamespaces is a list of namespace from which service accounts will receive this role. Either TargetNamespaceSelector or TargetNamespaces can be specified.
+	// +kubebuilder:validation:Optional
+	// +kubebuilder:validation:MinItems=1
+	// kubebuilder:validation:UniqueItems=true
+	// +listType=set
+	TargetNamespaces []string `json:"targetNamespaces,omitempty"`
 }
