@@ -22,6 +22,7 @@ limitations under the License.
 package v1alpha1
 
 import (
+	"github.com/redhat-cop/vault-config-operator/api/v1alpha1/utils"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
@@ -89,6 +90,11 @@ func (in *AuthEngineMountList) DeepCopyObject() runtime.Object {
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *AuthEngineMountSpec) DeepCopyInto(out *AuthEngineMountSpec) {
 	*out = *in
+	if in.Connection != nil {
+		in, out := &in.Connection, &out.Connection
+		*out = new(utils.VaultConnection)
+		(*in).DeepCopyInto(*out)
+	}
 	in.Authentication.DeepCopyInto(&out.Authentication)
 	in.AuthMount.DeepCopyInto(&out.AuthMount)
 }
@@ -208,6 +214,11 @@ func (in *DBSEConfig) DeepCopyInto(out *DBSEConfig) {
 			(*out)[key] = val
 		}
 	}
+	if in.RootPasswordRotation != nil {
+		in, out := &in.RootPasswordRotation, &out.RootPasswordRotation
+		*out = new(RootPasswordRotation)
+		**out = **in
+	}
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new DBSEConfig.
@@ -253,6 +264,36 @@ func (in *DBSERole) DeepCopy() *DBSERole {
 		return nil
 	}
 	out := new(DBSERole)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *DBSEStaticRole) DeepCopyInto(out *DBSEStaticRole) {
+	*out = *in
+	if in.RotationStatements != nil {
+		in, out := &in.RotationStatements, &out.RotationStatements
+		*out = make([]string, len(*in))
+		copy(*out, *in)
+	}
+	if in.PasswordCredentialConfig != nil {
+		in, out := &in.PasswordCredentialConfig, &out.PasswordCredentialConfig
+		*out = new(PasswordCredentialConfig)
+		**out = **in
+	}
+	if in.RSAPrivateKeyCredentialConfig != nil {
+		in, out := &in.RSAPrivateKeyCredentialConfig, &out.RSAPrivateKeyCredentialConfig
+		*out = new(RSAPrivateKeyCredentialConfig)
+		**out = **in
+	}
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new DBSEStaticRole.
+func (in *DBSEStaticRole) DeepCopy() *DBSEStaticRole {
+	if in == nil {
+		return nil
+	}
+	out := new(DBSEStaticRole)
 	in.DeepCopyInto(out)
 	return out
 }
@@ -319,6 +360,11 @@ func (in *DatabaseSecretEngineConfigList) DeepCopyObject() runtime.Object {
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *DatabaseSecretEngineConfigSpec) DeepCopyInto(out *DatabaseSecretEngineConfigSpec) {
 	*out = *in
+	if in.Connection != nil {
+		in, out := &in.Connection, &out.Connection
+		*out = new(utils.VaultConnection)
+		(*in).DeepCopyInto(*out)
+	}
 	in.Authentication.DeepCopyInto(&out.Authentication)
 	in.DBSEConfig.DeepCopyInto(&out.DBSEConfig)
 	in.RootCredentials.DeepCopyInto(&out.RootCredentials)
@@ -344,6 +390,7 @@ func (in *DatabaseSecretEngineConfigStatus) DeepCopyInto(out *DatabaseSecretEngi
 			(*in)[i].DeepCopyInto(&(*out)[i])
 		}
 	}
+	in.LastRootPasswordRotation.DeepCopyInto(&out.LastRootPasswordRotation)
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new DatabaseSecretEngineConfigStatus.
@@ -418,6 +465,11 @@ func (in *DatabaseSecretEngineRoleList) DeepCopyObject() runtime.Object {
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *DatabaseSecretEngineRoleSpec) DeepCopyInto(out *DatabaseSecretEngineRoleSpec) {
 	*out = *in
+	if in.Connection != nil {
+		in, out := &in.Connection, &out.Connection
+		*out = new(utils.VaultConnection)
+		(*in).DeepCopyInto(*out)
+	}
 	in.Authentication.DeepCopyInto(&out.Authentication)
 	in.DBSERole.DeepCopyInto(&out.DBSERole)
 }
@@ -450,6 +502,109 @@ func (in *DatabaseSecretEngineRoleStatus) DeepCopy() *DatabaseSecretEngineRoleSt
 		return nil
 	}
 	out := new(DatabaseSecretEngineRoleStatus)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *DatabaseSecretEngineStaticRole) DeepCopyInto(out *DatabaseSecretEngineStaticRole) {
+	*out = *in
+	out.TypeMeta = in.TypeMeta
+	in.ObjectMeta.DeepCopyInto(&out.ObjectMeta)
+	in.Spec.DeepCopyInto(&out.Spec)
+	in.Status.DeepCopyInto(&out.Status)
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new DatabaseSecretEngineStaticRole.
+func (in *DatabaseSecretEngineStaticRole) DeepCopy() *DatabaseSecretEngineStaticRole {
+	if in == nil {
+		return nil
+	}
+	out := new(DatabaseSecretEngineStaticRole)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyObject is an autogenerated deepcopy function, copying the receiver, creating a new runtime.Object.
+func (in *DatabaseSecretEngineStaticRole) DeepCopyObject() runtime.Object {
+	if c := in.DeepCopy(); c != nil {
+		return c
+	}
+	return nil
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *DatabaseSecretEngineStaticRoleList) DeepCopyInto(out *DatabaseSecretEngineStaticRoleList) {
+	*out = *in
+	out.TypeMeta = in.TypeMeta
+	in.ListMeta.DeepCopyInto(&out.ListMeta)
+	if in.Items != nil {
+		in, out := &in.Items, &out.Items
+		*out = make([]DatabaseSecretEngineStaticRole, len(*in))
+		for i := range *in {
+			(*in)[i].DeepCopyInto(&(*out)[i])
+		}
+	}
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new DatabaseSecretEngineStaticRoleList.
+func (in *DatabaseSecretEngineStaticRoleList) DeepCopy() *DatabaseSecretEngineStaticRoleList {
+	if in == nil {
+		return nil
+	}
+	out := new(DatabaseSecretEngineStaticRoleList)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyObject is an autogenerated deepcopy function, copying the receiver, creating a new runtime.Object.
+func (in *DatabaseSecretEngineStaticRoleList) DeepCopyObject() runtime.Object {
+	if c := in.DeepCopy(); c != nil {
+		return c
+	}
+	return nil
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *DatabaseSecretEngineStaticRoleSpec) DeepCopyInto(out *DatabaseSecretEngineStaticRoleSpec) {
+	*out = *in
+	if in.Connection != nil {
+		in, out := &in.Connection, &out.Connection
+		*out = new(utils.VaultConnection)
+		(*in).DeepCopyInto(*out)
+	}
+	in.Authentication.DeepCopyInto(&out.Authentication)
+	in.DBSEStaticRole.DeepCopyInto(&out.DBSEStaticRole)
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new DatabaseSecretEngineStaticRoleSpec.
+func (in *DatabaseSecretEngineStaticRoleSpec) DeepCopy() *DatabaseSecretEngineStaticRoleSpec {
+	if in == nil {
+		return nil
+	}
+	out := new(DatabaseSecretEngineStaticRoleSpec)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *DatabaseSecretEngineStaticRoleStatus) DeepCopyInto(out *DatabaseSecretEngineStaticRoleStatus) {
+	*out = *in
+	if in.Conditions != nil {
+		in, out := &in.Conditions, &out.Conditions
+		*out = make([]v1.Condition, len(*in))
+		for i := range *in {
+			(*in)[i].DeepCopyInto(&(*out)[i])
+		}
+	}
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new DatabaseSecretEngineStaticRoleStatus.
+func (in *DatabaseSecretEngineStaticRoleStatus) DeepCopy() *DatabaseSecretEngineStaticRoleStatus {
+	if in == nil {
+		return nil
+	}
+	out := new(DatabaseSecretEngineStaticRoleStatus)
 	in.DeepCopyInto(out)
 	return out
 }
@@ -531,6 +686,11 @@ func (in *GitHubSecretEngineConfigList) DeepCopyObject() runtime.Object {
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *GitHubSecretEngineConfigSpec) DeepCopyInto(out *GitHubSecretEngineConfigSpec) {
 	*out = *in
+	if in.Connection != nil {
+		in, out := &in.Connection, &out.Connection
+		*out = new(utils.VaultConnection)
+		(*in).DeepCopyInto(*out)
+	}
 	in.Authentication.DeepCopyInto(&out.Authentication)
 	out.GHConfig = in.GHConfig
 	in.SSHKeyReference.DeepCopyInto(&out.SSHKeyReference)
@@ -630,6 +790,11 @@ func (in *GitHubSecretEngineRoleList) DeepCopyObject() runtime.Object {
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *GitHubSecretEngineRoleSpec) DeepCopyInto(out *GitHubSecretEngineRoleSpec) {
 	*out = *in
+	if in.Connection != nil {
+		in, out := &in.Connection, &out.Connection
+		*out = new(utils.VaultConnection)
+		(*in).DeepCopyInto(*out)
+	}
 	in.Authentication.DeepCopyInto(&out.Authentication)
 	in.PermissionSet.DeepCopyInto(&out.PermissionSet)
 }
@@ -667,6 +832,304 @@ func (in *GitHubSecretEngineRoleStatus) DeepCopy() *GitHubSecretEngineRoleStatus
 }
 
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *JWTOIDCAuthEngineConfig) DeepCopyInto(out *JWTOIDCAuthEngineConfig) {
+	*out = *in
+	out.TypeMeta = in.TypeMeta
+	in.ObjectMeta.DeepCopyInto(&out.ObjectMeta)
+	in.Spec.DeepCopyInto(&out.Spec)
+	in.Status.DeepCopyInto(&out.Status)
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new JWTOIDCAuthEngineConfig.
+func (in *JWTOIDCAuthEngineConfig) DeepCopy() *JWTOIDCAuthEngineConfig {
+	if in == nil {
+		return nil
+	}
+	out := new(JWTOIDCAuthEngineConfig)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyObject is an autogenerated deepcopy function, copying the receiver, creating a new runtime.Object.
+func (in *JWTOIDCAuthEngineConfig) DeepCopyObject() runtime.Object {
+	if c := in.DeepCopy(); c != nil {
+		return c
+	}
+	return nil
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *JWTOIDCAuthEngineConfigList) DeepCopyInto(out *JWTOIDCAuthEngineConfigList) {
+	*out = *in
+	out.TypeMeta = in.TypeMeta
+	in.ListMeta.DeepCopyInto(&out.ListMeta)
+	if in.Items != nil {
+		in, out := &in.Items, &out.Items
+		*out = make([]JWTOIDCAuthEngineConfig, len(*in))
+		for i := range *in {
+			(*in)[i].DeepCopyInto(&(*out)[i])
+		}
+	}
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new JWTOIDCAuthEngineConfigList.
+func (in *JWTOIDCAuthEngineConfigList) DeepCopy() *JWTOIDCAuthEngineConfigList {
+	if in == nil {
+		return nil
+	}
+	out := new(JWTOIDCAuthEngineConfigList)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyObject is an autogenerated deepcopy function, copying the receiver, creating a new runtime.Object.
+func (in *JWTOIDCAuthEngineConfigList) DeepCopyObject() runtime.Object {
+	if c := in.DeepCopy(); c != nil {
+		return c
+	}
+	return nil
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *JWTOIDCAuthEngineConfigSpec) DeepCopyInto(out *JWTOIDCAuthEngineConfigSpec) {
+	*out = *in
+	if in.Connection != nil {
+		in, out := &in.Connection, &out.Connection
+		*out = new(utils.VaultConnection)
+		(*in).DeepCopyInto(*out)
+	}
+	in.Authentication.DeepCopyInto(&out.Authentication)
+	in.JWTOIDCConfig.DeepCopyInto(&out.JWTOIDCConfig)
+	in.OIDCCredentials.DeepCopyInto(&out.OIDCCredentials)
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new JWTOIDCAuthEngineConfigSpec.
+func (in *JWTOIDCAuthEngineConfigSpec) DeepCopy() *JWTOIDCAuthEngineConfigSpec {
+	if in == nil {
+		return nil
+	}
+	out := new(JWTOIDCAuthEngineConfigSpec)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *JWTOIDCAuthEngineConfigStatus) DeepCopyInto(out *JWTOIDCAuthEngineConfigStatus) {
+	*out = *in
+	if in.Conditions != nil {
+		in, out := &in.Conditions, &out.Conditions
+		*out = make([]v1.Condition, len(*in))
+		for i := range *in {
+			(*in)[i].DeepCopyInto(&(*out)[i])
+		}
+	}
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new JWTOIDCAuthEngineConfigStatus.
+func (in *JWTOIDCAuthEngineConfigStatus) DeepCopy() *JWTOIDCAuthEngineConfigStatus {
+	if in == nil {
+		return nil
+	}
+	out := new(JWTOIDCAuthEngineConfigStatus)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *JWTOIDCAuthEngineRole) DeepCopyInto(out *JWTOIDCAuthEngineRole) {
+	*out = *in
+	out.TypeMeta = in.TypeMeta
+	in.ObjectMeta.DeepCopyInto(&out.ObjectMeta)
+	in.Spec.DeepCopyInto(&out.Spec)
+	in.Status.DeepCopyInto(&out.Status)
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new JWTOIDCAuthEngineRole.
+func (in *JWTOIDCAuthEngineRole) DeepCopy() *JWTOIDCAuthEngineRole {
+	if in == nil {
+		return nil
+	}
+	out := new(JWTOIDCAuthEngineRole)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyObject is an autogenerated deepcopy function, copying the receiver, creating a new runtime.Object.
+func (in *JWTOIDCAuthEngineRole) DeepCopyObject() runtime.Object {
+	if c := in.DeepCopy(); c != nil {
+		return c
+	}
+	return nil
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *JWTOIDCAuthEngineRoleList) DeepCopyInto(out *JWTOIDCAuthEngineRoleList) {
+	*out = *in
+	out.TypeMeta = in.TypeMeta
+	in.ListMeta.DeepCopyInto(&out.ListMeta)
+	if in.Items != nil {
+		in, out := &in.Items, &out.Items
+		*out = make([]JWTOIDCAuthEngineRole, len(*in))
+		for i := range *in {
+			(*in)[i].DeepCopyInto(&(*out)[i])
+		}
+	}
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new JWTOIDCAuthEngineRoleList.
+func (in *JWTOIDCAuthEngineRoleList) DeepCopy() *JWTOIDCAuthEngineRoleList {
+	if in == nil {
+		return nil
+	}
+	out := new(JWTOIDCAuthEngineRoleList)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyObject is an autogenerated deepcopy function, copying the receiver, creating a new runtime.Object.
+func (in *JWTOIDCAuthEngineRoleList) DeepCopyObject() runtime.Object {
+	if c := in.DeepCopy(); c != nil {
+		return c
+	}
+	return nil
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *JWTOIDCAuthEngineRoleSpec) DeepCopyInto(out *JWTOIDCAuthEngineRoleSpec) {
+	*out = *in
+	if in.Connection != nil {
+		in, out := &in.Connection, &out.Connection
+		*out = new(utils.VaultConnection)
+		(*in).DeepCopyInto(*out)
+	}
+	in.Authentication.DeepCopyInto(&out.Authentication)
+	in.JWTOIDCRole.DeepCopyInto(&out.JWTOIDCRole)
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new JWTOIDCAuthEngineRoleSpec.
+func (in *JWTOIDCAuthEngineRoleSpec) DeepCopy() *JWTOIDCAuthEngineRoleSpec {
+	if in == nil {
+		return nil
+	}
+	out := new(JWTOIDCAuthEngineRoleSpec)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *JWTOIDCAuthEngineRoleStatus) DeepCopyInto(out *JWTOIDCAuthEngineRoleStatus) {
+	*out = *in
+	if in.Conditions != nil {
+		in, out := &in.Conditions, &out.Conditions
+		*out = make([]v1.Condition, len(*in))
+		for i := range *in {
+			(*in)[i].DeepCopyInto(&(*out)[i])
+		}
+	}
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new JWTOIDCAuthEngineRoleStatus.
+func (in *JWTOIDCAuthEngineRoleStatus) DeepCopy() *JWTOIDCAuthEngineRoleStatus {
+	if in == nil {
+		return nil
+	}
+	out := new(JWTOIDCAuthEngineRoleStatus)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *JWTOIDCConfig) DeepCopyInto(out *JWTOIDCConfig) {
+	*out = *in
+	if in.OIDCResponseTypes != nil {
+		in, out := &in.OIDCResponseTypes, &out.OIDCResponseTypes
+		*out = make([]string, len(*in))
+		copy(*out, *in)
+	}
+	if in.JWTValidationPubKeys != nil {
+		in, out := &in.JWTValidationPubKeys, &out.JWTValidationPubKeys
+		*out = make([]string, len(*in))
+		copy(*out, *in)
+	}
+	if in.JWTSupportedAlgs != nil {
+		in, out := &in.JWTSupportedAlgs, &out.JWTSupportedAlgs
+		*out = make([]string, len(*in))
+		copy(*out, *in)
+	}
+	if in.ProviderConfig != nil {
+		in, out := &in.ProviderConfig, &out.ProviderConfig
+		*out = make(map[string]string, len(*in))
+		for key, val := range *in {
+			(*out)[key] = val
+		}
+	}
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new JWTOIDCConfig.
+func (in *JWTOIDCConfig) DeepCopy() *JWTOIDCConfig {
+	if in == nil {
+		return nil
+	}
+	out := new(JWTOIDCConfig)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *JWTOIDCRole) DeepCopyInto(out *JWTOIDCRole) {
+	*out = *in
+	if in.BoundAudiences != nil {
+		in, out := &in.BoundAudiences, &out.BoundAudiences
+		*out = make([]string, len(*in))
+		copy(*out, *in)
+	}
+	if in.BoundClaims != nil {
+		in, out := &in.BoundClaims, &out.BoundClaims
+		*out = make(map[string]string, len(*in))
+		for key, val := range *in {
+			(*out)[key] = val
+		}
+	}
+	if in.ClaimMappings != nil {
+		in, out := &in.ClaimMappings, &out.ClaimMappings
+		*out = make(map[string]string, len(*in))
+		for key, val := range *in {
+			(*out)[key] = val
+		}
+	}
+	if in.OIDCScopes != nil {
+		in, out := &in.OIDCScopes, &out.OIDCScopes
+		*out = make([]string, len(*in))
+		copy(*out, *in)
+	}
+	if in.AllowedRedirectURIs != nil {
+		in, out := &in.AllowedRedirectURIs, &out.AllowedRedirectURIs
+		*out = make([]string, len(*in))
+		copy(*out, *in)
+	}
+	if in.TokenPolicies != nil {
+		in, out := &in.TokenPolicies, &out.TokenPolicies
+		*out = make([]string, len(*in))
+		copy(*out, *in)
+	}
+	if in.TokenBoundCIDRs != nil {
+		in, out := &in.TokenBoundCIDRs, &out.TokenBoundCIDRs
+		*out = make([]string, len(*in))
+		copy(*out, *in)
+	}
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new JWTOIDCRole.
+func (in *JWTOIDCRole) DeepCopy() *JWTOIDCRole {
+	if in == nil {
+		return nil
+	}
+	out := new(JWTOIDCRole)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *KAECConfig) DeepCopyInto(out *KAECConfig) {
 	*out = *in
 	if in.PEMKeys != nil {
@@ -687,21 +1150,52 @@ func (in *KAECConfig) DeepCopy() *KAECConfig {
 }
 
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
-func (in *KubeAuthConfiguration) DeepCopyInto(out *KubeAuthConfiguration) {
+func (in *KubeSEConfig) DeepCopyInto(out *KubeSEConfig) {
 	*out = *in
-	if in.ServiceAccount != nil {
-		in, out := &in.ServiceAccount, &out.ServiceAccount
-		*out = new(corev1.LocalObjectReference)
-		**out = **in
-	}
 }
 
-// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new KubeAuthConfiguration.
-func (in *KubeAuthConfiguration) DeepCopy() *KubeAuthConfiguration {
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new KubeSEConfig.
+func (in *KubeSEConfig) DeepCopy() *KubeSEConfig {
 	if in == nil {
 		return nil
 	}
-	out := new(KubeAuthConfiguration)
+	out := new(KubeSEConfig)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *KubeSERole) DeepCopyInto(out *KubeSERole) {
+	*out = *in
+	if in.AllowedKubernetesNamespaces != nil {
+		in, out := &in.AllowedKubernetesNamespaces, &out.AllowedKubernetesNamespaces
+		*out = make([]string, len(*in))
+		copy(*out, *in)
+	}
+	out.DefaultTTL = in.DefaultTTL
+	out.MaxTTL = in.MaxTTL
+	if in.ExtraAnnotations != nil {
+		in, out := &in.ExtraAnnotations, &out.ExtraAnnotations
+		*out = make(map[string]string, len(*in))
+		for key, val := range *in {
+			(*out)[key] = val
+		}
+	}
+	if in.ExtraLabels != nil {
+		in, out := &in.ExtraLabels, &out.ExtraLabels
+		*out = make(map[string]string, len(*in))
+		for key, val := range *in {
+			(*out)[key] = val
+		}
+	}
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new KubeSERole.
+func (in *KubeSERole) DeepCopy() *KubeSERole {
+	if in == nil {
+		return nil
+	}
+	out := new(KubeSERole)
 	in.DeepCopyInto(out)
 	return out
 }
@@ -768,6 +1262,11 @@ func (in *KubernetesAuthEngineConfigList) DeepCopyObject() runtime.Object {
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *KubernetesAuthEngineConfigSpec) DeepCopyInto(out *KubernetesAuthEngineConfigSpec) {
 	*out = *in
+	if in.Connection != nil {
+		in, out := &in.Connection, &out.Connection
+		*out = new(utils.VaultConnection)
+		(*in).DeepCopyInto(*out)
+	}
 	in.Authentication.DeepCopyInto(&out.Authentication)
 	in.KAECConfig.DeepCopyInto(&out.KAECConfig)
 	if in.TokenReviewerServiceAccount != nil {
@@ -871,6 +1370,11 @@ func (in *KubernetesAuthEngineRoleList) DeepCopyObject() runtime.Object {
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *KubernetesAuthEngineRoleSpec) DeepCopyInto(out *KubernetesAuthEngineRoleSpec) {
 	*out = *in
+	if in.Connection != nil {
+		in, out := &in.Connection, &out.Connection
+		*out = new(utils.VaultConnection)
+		(*in).DeepCopyInto(*out)
+	}
 	in.Authentication.DeepCopyInto(&out.Authentication)
 	in.VRole.DeepCopyInto(&out.VRole)
 	in.TargetNamespaces.DeepCopyInto(&out.TargetNamespaces)
@@ -904,6 +1408,214 @@ func (in *KubernetesAuthEngineRoleStatus) DeepCopy() *KubernetesAuthEngineRoleSt
 		return nil
 	}
 	out := new(KubernetesAuthEngineRoleStatus)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *KubernetesSecretEngineConfig) DeepCopyInto(out *KubernetesSecretEngineConfig) {
+	*out = *in
+	out.TypeMeta = in.TypeMeta
+	in.ObjectMeta.DeepCopyInto(&out.ObjectMeta)
+	in.Spec.DeepCopyInto(&out.Spec)
+	in.Status.DeepCopyInto(&out.Status)
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new KubernetesSecretEngineConfig.
+func (in *KubernetesSecretEngineConfig) DeepCopy() *KubernetesSecretEngineConfig {
+	if in == nil {
+		return nil
+	}
+	out := new(KubernetesSecretEngineConfig)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyObject is an autogenerated deepcopy function, copying the receiver, creating a new runtime.Object.
+func (in *KubernetesSecretEngineConfig) DeepCopyObject() runtime.Object {
+	if c := in.DeepCopy(); c != nil {
+		return c
+	}
+	return nil
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *KubernetesSecretEngineConfigList) DeepCopyInto(out *KubernetesSecretEngineConfigList) {
+	*out = *in
+	out.TypeMeta = in.TypeMeta
+	in.ListMeta.DeepCopyInto(&out.ListMeta)
+	if in.Items != nil {
+		in, out := &in.Items, &out.Items
+		*out = make([]KubernetesSecretEngineConfig, len(*in))
+		for i := range *in {
+			(*in)[i].DeepCopyInto(&(*out)[i])
+		}
+	}
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new KubernetesSecretEngineConfigList.
+func (in *KubernetesSecretEngineConfigList) DeepCopy() *KubernetesSecretEngineConfigList {
+	if in == nil {
+		return nil
+	}
+	out := new(KubernetesSecretEngineConfigList)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyObject is an autogenerated deepcopy function, copying the receiver, creating a new runtime.Object.
+func (in *KubernetesSecretEngineConfigList) DeepCopyObject() runtime.Object {
+	if c := in.DeepCopy(); c != nil {
+		return c
+	}
+	return nil
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *KubernetesSecretEngineConfigSpec) DeepCopyInto(out *KubernetesSecretEngineConfigSpec) {
+	*out = *in
+	if in.Connection != nil {
+		in, out := &in.Connection, &out.Connection
+		*out = new(utils.VaultConnection)
+		(*in).DeepCopyInto(*out)
+	}
+	in.Authentication.DeepCopyInto(&out.Authentication)
+	in.JWTReference.DeepCopyInto(&out.JWTReference)
+	out.KubeSEConfig = in.KubeSEConfig
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new KubernetesSecretEngineConfigSpec.
+func (in *KubernetesSecretEngineConfigSpec) DeepCopy() *KubernetesSecretEngineConfigSpec {
+	if in == nil {
+		return nil
+	}
+	out := new(KubernetesSecretEngineConfigSpec)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *KubernetesSecretEngineConfigStatus) DeepCopyInto(out *KubernetesSecretEngineConfigStatus) {
+	*out = *in
+	if in.Conditions != nil {
+		in, out := &in.Conditions, &out.Conditions
+		*out = make([]v1.Condition, len(*in))
+		for i := range *in {
+			(*in)[i].DeepCopyInto(&(*out)[i])
+		}
+	}
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new KubernetesSecretEngineConfigStatus.
+func (in *KubernetesSecretEngineConfigStatus) DeepCopy() *KubernetesSecretEngineConfigStatus {
+	if in == nil {
+		return nil
+	}
+	out := new(KubernetesSecretEngineConfigStatus)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *KubernetesSecretEngineRole) DeepCopyInto(out *KubernetesSecretEngineRole) {
+	*out = *in
+	out.TypeMeta = in.TypeMeta
+	in.ObjectMeta.DeepCopyInto(&out.ObjectMeta)
+	in.Spec.DeepCopyInto(&out.Spec)
+	in.Status.DeepCopyInto(&out.Status)
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new KubernetesSecretEngineRole.
+func (in *KubernetesSecretEngineRole) DeepCopy() *KubernetesSecretEngineRole {
+	if in == nil {
+		return nil
+	}
+	out := new(KubernetesSecretEngineRole)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyObject is an autogenerated deepcopy function, copying the receiver, creating a new runtime.Object.
+func (in *KubernetesSecretEngineRole) DeepCopyObject() runtime.Object {
+	if c := in.DeepCopy(); c != nil {
+		return c
+	}
+	return nil
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *KubernetesSecretEngineRoleList) DeepCopyInto(out *KubernetesSecretEngineRoleList) {
+	*out = *in
+	out.TypeMeta = in.TypeMeta
+	in.ListMeta.DeepCopyInto(&out.ListMeta)
+	if in.Items != nil {
+		in, out := &in.Items, &out.Items
+		*out = make([]KubernetesSecretEngineRole, len(*in))
+		for i := range *in {
+			(*in)[i].DeepCopyInto(&(*out)[i])
+		}
+	}
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new KubernetesSecretEngineRoleList.
+func (in *KubernetesSecretEngineRoleList) DeepCopy() *KubernetesSecretEngineRoleList {
+	if in == nil {
+		return nil
+	}
+	out := new(KubernetesSecretEngineRoleList)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyObject is an autogenerated deepcopy function, copying the receiver, creating a new runtime.Object.
+func (in *KubernetesSecretEngineRoleList) DeepCopyObject() runtime.Object {
+	if c := in.DeepCopy(); c != nil {
+		return c
+	}
+	return nil
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *KubernetesSecretEngineRoleSpec) DeepCopyInto(out *KubernetesSecretEngineRoleSpec) {
+	*out = *in
+	if in.Connection != nil {
+		in, out := &in.Connection, &out.Connection
+		*out = new(utils.VaultConnection)
+		(*in).DeepCopyInto(*out)
+	}
+	in.Authentication.DeepCopyInto(&out.Authentication)
+	in.TargetNamespaces.DeepCopyInto(&out.TargetNamespaces)
+	in.KubeSERole.DeepCopyInto(&out.KubeSERole)
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new KubernetesSecretEngineRoleSpec.
+func (in *KubernetesSecretEngineRoleSpec) DeepCopy() *KubernetesSecretEngineRoleSpec {
+	if in == nil {
+		return nil
+	}
+	out := new(KubernetesSecretEngineRoleSpec)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *KubernetesSecretEngineRoleStatus) DeepCopyInto(out *KubernetesSecretEngineRoleStatus) {
+	*out = *in
+	if in.Conditions != nil {
+		in, out := &in.Conditions, &out.Conditions
+		*out = make([]v1.Condition, len(*in))
+		for i := range *in {
+			(*in)[i].DeepCopyInto(&(*out)[i])
+		}
+	}
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new KubernetesSecretEngineRoleStatus.
+func (in *KubernetesSecretEngineRoleStatus) DeepCopy() *KubernetesSecretEngineRoleStatus {
+	if in == nil {
+		return nil
+	}
+	out := new(KubernetesSecretEngineRoleStatus)
 	in.DeepCopyInto(out)
 	return out
 }
@@ -970,6 +1682,11 @@ func (in *LDAPAuthEngineConfigList) DeepCopyObject() runtime.Object {
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *LDAPAuthEngineConfigSpec) DeepCopyInto(out *LDAPAuthEngineConfigSpec) {
 	*out = *in
+	if in.Connection != nil {
+		in, out := &in.Connection, &out.Connection
+		*out = new(utils.VaultConnection)
+		(*in).DeepCopyInto(*out)
+	}
 	in.Authentication.DeepCopyInto(&out.Authentication)
 	out.LDAPConfig = in.LDAPConfig
 	in.BindCredentials.DeepCopyInto(&out.BindCredentials)
@@ -1069,6 +1786,11 @@ func (in *LDAPAuthEngineGroupList) DeepCopyObject() runtime.Object {
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *LDAPAuthEngineGroupSpec) DeepCopyInto(out *LDAPAuthEngineGroupSpec) {
 	*out = *in
+	if in.Connection != nil {
+		in, out := &in.Connection, &out.Connection
+		*out = new(utils.VaultConnection)
+		(*in).DeepCopyInto(*out)
+	}
 	in.Authentication.DeepCopyInto(&out.Authentication)
 }
 
@@ -1396,6 +2118,11 @@ func (in *PKISecretEngineConfigList) DeepCopyObject() runtime.Object {
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *PKISecretEngineConfigSpec) DeepCopyInto(out *PKISecretEngineConfigSpec) {
 	*out = *in
+	if in.Connection != nil {
+		in, out := &in.Connection, &out.Connection
+		*out = new(utils.VaultConnection)
+		(*in).DeepCopyInto(*out)
+	}
 	in.Authentication.DeepCopyInto(&out.Authentication)
 	out.PKIType = in.PKIType
 	in.PKICommon.DeepCopyInto(&out.PKICommon)
@@ -1497,6 +2224,11 @@ func (in *PKISecretEngineRoleList) DeepCopyObject() runtime.Object {
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *PKISecretEngineRoleSpec) DeepCopyInto(out *PKISecretEngineRoleSpec) {
 	*out = *in
+	if in.Connection != nil {
+		in, out := &in.Connection, &out.Connection
+		*out = new(utils.VaultConnection)
+		(*in).DeepCopyInto(*out)
+	}
 	in.Authentication.DeepCopyInto(&out.Authentication)
 	in.PKIRole.DeepCopyInto(&out.PKIRole)
 }
@@ -1544,6 +2276,21 @@ func (in *PKIType) DeepCopy() *PKIType {
 		return nil
 	}
 	out := new(PKIType)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *PasswordCredentialConfig) DeepCopyInto(out *PasswordCredentialConfig) {
+	*out = *in
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new PasswordCredentialConfig.
+func (in *PasswordCredentialConfig) DeepCopy() *PasswordCredentialConfig {
+	if in == nil {
+		return nil
+	}
+	out := new(PasswordCredentialConfig)
 	in.DeepCopyInto(out)
 	return out
 }
@@ -1645,6 +2392,11 @@ func (in *PasswordPolicyRule) DeepCopy() *PasswordPolicyRule {
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *PasswordPolicySpec) DeepCopyInto(out *PasswordPolicySpec) {
 	*out = *in
+	if in.Connection != nil {
+		in, out := &in.Connection, &out.Connection
+		*out = new(utils.VaultConnection)
+		(*in).DeepCopyInto(*out)
+	}
 	in.Authentication.DeepCopyInto(&out.Authentication)
 }
 
@@ -1774,6 +2526,11 @@ func (in *PolicyList) DeepCopyObject() runtime.Object {
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *PolicySpec) DeepCopyInto(out *PolicySpec) {
 	*out = *in
+	if in.Connection != nil {
+		in, out := &in.Connection, &out.Connection
+		*out = new(utils.VaultConnection)
+		(*in).DeepCopyInto(*out)
+	}
 	in.Authentication.DeepCopyInto(&out.Authentication)
 }
 
@@ -1959,6 +2716,11 @@ func (in *QuaySecretEngineConfigList) DeepCopyObject() runtime.Object {
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *QuaySecretEngineConfigSpec) DeepCopyInto(out *QuaySecretEngineConfigSpec) {
 	*out = *in
+	if in.Connection != nil {
+		in, out := &in.Connection, &out.Connection
+		*out = new(utils.VaultConnection)
+		(*in).DeepCopyInto(*out)
+	}
 	in.Authentication.DeepCopyInto(&out.Authentication)
 	out.QuayConfig = in.QuayConfig
 	in.RootCredentials.DeepCopyInto(&out.RootCredentials)
@@ -2058,6 +2820,11 @@ func (in *QuaySecretEngineRoleList) DeepCopyObject() runtime.Object {
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *QuaySecretEngineRoleSpec) DeepCopyInto(out *QuaySecretEngineRoleSpec) {
 	*out = *in
+	if in.Connection != nil {
+		in, out := &in.Connection, &out.Connection
+		*out = new(utils.VaultConnection)
+		(*in).DeepCopyInto(*out)
+	}
 	in.Authentication.DeepCopyInto(&out.Authentication)
 	in.QuayRole.DeepCopyInto(&out.QuayRole)
 }
@@ -2156,6 +2923,11 @@ func (in *QuaySecretEngineStaticRoleList) DeepCopyObject() runtime.Object {
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *QuaySecretEngineStaticRoleSpec) DeepCopyInto(out *QuaySecretEngineStaticRoleSpec) {
 	*out = *in
+	if in.Connection != nil {
+		in, out := &in.Connection, &out.Connection
+		*out = new(utils.VaultConnection)
+		(*in).DeepCopyInto(*out)
+	}
 	in.Authentication.DeepCopyInto(&out.Authentication)
 	in.QuayBaseRole.DeepCopyInto(&out.QuayBaseRole)
 }
@@ -2235,6 +3007,21 @@ func (in *RMQSERole) DeepCopy() *RMQSERole {
 }
 
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *RSAPrivateKeyCredentialConfig) DeepCopyInto(out *RSAPrivateKeyCredentialConfig) {
+	*out = *in
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new RSAPrivateKeyCredentialConfig.
+func (in *RSAPrivateKeyCredentialConfig) DeepCopy() *RSAPrivateKeyCredentialConfig {
+	if in == nil {
+		return nil
+	}
+	out := new(RSAPrivateKeyCredentialConfig)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *RabbitMQSecretEngineConfig) DeepCopyInto(out *RabbitMQSecretEngineConfig) {
 	*out = *in
 	out.TypeMeta = in.TypeMeta
@@ -2296,6 +3083,11 @@ func (in *RabbitMQSecretEngineConfigList) DeepCopyObject() runtime.Object {
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *RabbitMQSecretEngineConfigSpec) DeepCopyInto(out *RabbitMQSecretEngineConfigSpec) {
 	*out = *in
+	if in.Connection != nil {
+		in, out := &in.Connection, &out.Connection
+		*out = new(utils.VaultConnection)
+		(*in).DeepCopyInto(*out)
+	}
 	in.Authentication.DeepCopyInto(&out.Authentication)
 	out.RMQSEConfig = in.RMQSEConfig
 	in.RootCredentials.DeepCopyInto(&out.RootCredentials)
@@ -2395,6 +3187,11 @@ func (in *RabbitMQSecretEngineRoleList) DeepCopyObject() runtime.Object {
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *RabbitMQSecretEngineRoleSpec) DeepCopyInto(out *RabbitMQSecretEngineRoleSpec) {
 	*out = *in
+	if in.Connection != nil {
+		in, out := &in.Connection, &out.Connection
+		*out = new(utils.VaultConnection)
+		(*in).DeepCopyInto(*out)
+	}
 	in.Authentication.DeepCopyInto(&out.Authentication)
 	in.RMQSERole.DeepCopyInto(&out.RMQSERole)
 }
@@ -2493,6 +3290,11 @@ func (in *RandomSecretList) DeepCopyObject() runtime.Object {
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *RandomSecretSpec) DeepCopyInto(out *RandomSecretSpec) {
 	*out = *in
+	if in.Connection != nil {
+		in, out := &in.Connection, &out.Connection
+		*out = new(utils.VaultConnection)
+		(*in).DeepCopyInto(*out)
+	}
 	in.Authentication.DeepCopyInto(&out.Authentication)
 	out.SecretFormat = in.SecretFormat
 	if in.RefreshPeriod != nil {
@@ -2539,31 +3341,17 @@ func (in *RandomSecretStatus) DeepCopy() *RandomSecretStatus {
 }
 
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
-func (in *RootCredentialConfig) DeepCopyInto(out *RootCredentialConfig) {
+func (in *RootPasswordRotation) DeepCopyInto(out *RootPasswordRotation) {
 	*out = *in
-	if in.VaultSecret != nil {
-		in, out := &in.VaultSecret, &out.VaultSecret
-		*out = new(VaultSecretReference)
-		**out = **in
-	}
-	if in.Secret != nil {
-		in, out := &in.Secret, &out.Secret
-		*out = new(corev1.LocalObjectReference)
-		**out = **in
-	}
-	if in.RandomSecret != nil {
-		in, out := &in.RandomSecret, &out.RandomSecret
-		*out = new(corev1.LocalObjectReference)
-		**out = **in
-	}
+	out.RotationPeriod = in.RotationPeriod
 }
 
-// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new RootCredentialConfig.
-func (in *RootCredentialConfig) DeepCopy() *RootCredentialConfig {
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new RootPasswordRotation.
+func (in *RootPasswordRotation) DeepCopy() *RootPasswordRotation {
 	if in == nil {
 		return nil
 	}
-	out := new(RootCredentialConfig)
+	out := new(RootPasswordRotation)
 	in.DeepCopyInto(out)
 	return out
 }
@@ -2573,7 +3361,7 @@ func (in *SSHKeyConfig) DeepCopyInto(out *SSHKeyConfig) {
 	*out = *in
 	if in.VaultSecret != nil {
 		in, out := &in.VaultSecret, &out.VaultSecret
-		*out = new(VaultSecretReference)
+		*out = new(utils.VaultSecretReference)
 		**out = **in
 	}
 	if in.Secret != nil {
@@ -2655,6 +3443,11 @@ func (in *SecretEngineMountList) DeepCopyObject() runtime.Object {
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *SecretEngineMountSpec) DeepCopyInto(out *SecretEngineMountSpec) {
 	*out = *in
+	if in.Connection != nil {
+		in, out := &in.Connection, &out.Connection
+		*out = new(utils.VaultConnection)
+		(*in).DeepCopyInto(*out)
+	}
 	in.Authentication.DeepCopyInto(&out.Authentication)
 	in.Mount.DeepCopyInto(&out.Mount)
 }
@@ -2687,31 +3480,6 @@ func (in *SecretEngineMountStatus) DeepCopy() *SecretEngineMountStatus {
 		return nil
 	}
 	out := new(SecretEngineMountStatus)
-	in.DeepCopyInto(out)
-	return out
-}
-
-// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
-func (in *TargetNamespaceConfig) DeepCopyInto(out *TargetNamespaceConfig) {
-	*out = *in
-	if in.TargetNamespaceSelector != nil {
-		in, out := &in.TargetNamespaceSelector, &out.TargetNamespaceSelector
-		*out = new(v1.LabelSelector)
-		(*in).DeepCopyInto(*out)
-	}
-	if in.TargetNamespaces != nil {
-		in, out := &in.TargetNamespaces, &out.TargetNamespaces
-		*out = make([]string, len(*in))
-		copy(*out, *in)
-	}
-}
-
-// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new TargetNamespaceConfig.
-func (in *TargetNamespaceConfig) DeepCopy() *TargetNamespaceConfig {
-	if in == nil {
-		return nil
-	}
-	out := new(TargetNamespaceConfig)
 	in.DeepCopyInto(out)
 	return out
 }
@@ -2775,6 +3543,11 @@ func (in *VRole) DeepCopyInto(out *VRole) {
 		in, out := &in.TargetServiceAccounts, &out.TargetServiceAccounts
 		*out = make([]string, len(*in))
 		copy(*out, *in)
+	}
+	if in.Audience != nil {
+		in, out := &in.Audience, &out.Audience
+		*out = new(string)
+		**out = **in
 	}
 	if in.Policies != nil {
 		in, out := &in.Policies, &out.Policies
@@ -2849,6 +3622,13 @@ func (in *VaultSecret) DeepCopyObject() runtime.Object {
 func (in *VaultSecretDefinition) DeepCopyInto(out *VaultSecretDefinition) {
 	*out = *in
 	in.Authentication.DeepCopyInto(&out.Authentication)
+	if in.RequestPayload != nil {
+		in, out := &in.RequestPayload, &out.RequestPayload
+		*out = make(map[string]string, len(*in))
+		for key, val := range *in {
+			(*out)[key] = val
+		}
+	}
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new VaultSecretDefinition.
@@ -2906,21 +3686,6 @@ func (in *VaultSecretList) DeepCopyObject() runtime.Object {
 		return c
 	}
 	return nil
-}
-
-// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
-func (in *VaultSecretReference) DeepCopyInto(out *VaultSecretReference) {
-	*out = *in
-}
-
-// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new VaultSecretReference.
-func (in *VaultSecretReference) DeepCopy() *VaultSecretReference {
-	if in == nil {
-		return nil
-	}
-	out := new(VaultSecretReference)
-	in.DeepCopyInto(out)
-	return out
 }
 
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
