@@ -6,9 +6,11 @@ package controllers
 import (
 	"time"
 
-	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	redhatcopv1alpha1 "github.com/redhat-cop/vault-config-operator/api/v1alpha1"
+	"github.com/redhat-cop/vault-config-operator/controllers/vaultresourcecontroller"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"k8s.io/apimachinery/pkg/types"
 )
@@ -36,7 +38,7 @@ var _ = Describe("PKISecretEngineConfig controller", func() {
 				}
 
 				for _, condition := range pCreated.Status.Conditions {
-					if condition.Type == "ReconcileSuccess" {
+					if condition.Type == vaultresourcecontroller.ReconcileSuccessful && condition.Status == metav1.ConditionTrue {
 						return true
 					}
 				}
@@ -59,7 +61,7 @@ var _ = Describe("PKISecretEngineConfig controller", func() {
 				}
 
 				for _, condition := range kaerCreated.Status.Conditions {
-					if condition.Type == "ReconcileSuccess" {
+					if condition.Type == vaultresourcecontroller.ReconcileSuccessful && condition.Status == metav1.ConditionTrue {
 						return true
 					}
 				}
@@ -84,7 +86,7 @@ var _ = Describe("PKISecretEngineConfig controller", func() {
 				}
 
 				for _, condition := range semCreated.Status.Conditions {
-					if condition.Type == "ReconcileSuccess" {
+					if condition.Type == vaultresourcecontroller.ReconcileSuccessful && condition.Status == metav1.ConditionTrue {
 						return true
 					}
 				}
@@ -112,7 +114,7 @@ var _ = Describe("PKISecretEngineConfig controller", func() {
 				}
 
 				for _, condition := range rsCreated.Status.Conditions {
-					if condition.Type == "ReconcileSuccess" {
+					if condition.Type == vaultresourcecontroller.ReconcileSuccessful && condition.Status == metav1.ConditionTrue {
 						return true
 					}
 				}
@@ -140,7 +142,7 @@ var _ = Describe("PKISecretEngineConfig controller", func() {
 				}
 
 				for _, condition := range rsCreated.Status.Conditions {
-					if condition.Type == "ReconcileSuccess" {
+					if condition.Type == vaultresourcecontroller.ReconcileSuccessful && condition.Status == metav1.ConditionTrue {
 						return true
 					}
 				}
