@@ -11,11 +11,13 @@ import (
 	"strings"
 	"time"
 
-	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	redhatcopv1alpha1 "github.com/redhat-cop/vault-config-operator/api/v1alpha1"
+	"github.com/redhat-cop/vault-config-operator/controllers/vaultresourcecontroller"
 
 	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 )
 
@@ -42,7 +44,7 @@ var _ = Describe("VaultSecret controller", func() {
 				}
 
 				for _, condition := range ppCreated.Status.Conditions {
-					if condition.Type == "ReconcileSuccess" {
+					if condition.Type == vaultresourcecontroller.ReconcileSuccessful && condition.Status == metav1.ConditionTrue {
 						return true
 					}
 				}
@@ -68,7 +70,7 @@ var _ = Describe("VaultSecret controller", func() {
 				}
 
 				for _, condition := range pCreated.Status.Conditions {
-					if condition.Type == "ReconcileSuccess" {
+					if condition.Type == vaultresourcecontroller.ReconcileSuccessful && condition.Status == metav1.ConditionTrue {
 						return true
 					}
 				}
@@ -93,7 +95,7 @@ var _ = Describe("VaultSecret controller", func() {
 				}
 
 				for _, condition := range pCreated.Status.Conditions {
-					if condition.Type == "ReconcileSuccess" {
+					if condition.Type == vaultresourcecontroller.ReconcileSuccessful && condition.Status == metav1.ConditionTrue {
 						return true
 					}
 				}
@@ -118,7 +120,7 @@ var _ = Describe("VaultSecret controller", func() {
 				}
 
 				for _, condition := range pCreated.Status.Conditions {
-					if condition.Type == "ReconcileSuccess" {
+					if condition.Type == vaultresourcecontroller.ReconcileSuccessful && condition.Status == metav1.ConditionTrue {
 						return true
 					}
 				}
@@ -143,7 +145,7 @@ var _ = Describe("VaultSecret controller", func() {
 				}
 
 				for _, condition := range kaerCreated.Status.Conditions {
-					if condition.Type == "ReconcileSuccess" {
+					if condition.Type == vaultresourcecontroller.ReconcileSuccessful && condition.Status == metav1.ConditionTrue {
 						return true
 					}
 				}
@@ -166,7 +168,7 @@ var _ = Describe("VaultSecret controller", func() {
 				}
 
 				for _, condition := range kaerCreated.Status.Conditions {
-					if condition.Type == "ReconcileSuccess" {
+					if condition.Type == vaultresourcecontroller.ReconcileSuccessful && condition.Status == metav1.ConditionTrue {
 						return true
 					}
 				}
@@ -189,7 +191,7 @@ var _ = Describe("VaultSecret controller", func() {
 				}
 
 				for _, condition := range kaerCreated.Status.Conditions {
-					if condition.Type == "ReconcileSuccess" {
+					if condition.Type == vaultresourcecontroller.ReconcileSuccessful && condition.Status == metav1.ConditionTrue {
 						return true
 					}
 				}
@@ -214,7 +216,7 @@ var _ = Describe("VaultSecret controller", func() {
 				}
 
 				for _, condition := range semCreated.Status.Conditions {
-					if condition.Type == "ReconcileSuccess" {
+					if condition.Type == vaultresourcecontroller.ReconcileSuccessful && condition.Status == metav1.ConditionTrue {
 						return true
 					}
 				}
@@ -239,7 +241,7 @@ var _ = Describe("VaultSecret controller", func() {
 				}
 
 				for _, condition := range rsCreated.Status.Conditions {
-					if condition.Type == "ReconcileSuccess" {
+					if condition.Type == vaultresourcecontroller.ReconcileSuccessful && condition.Status == metav1.ConditionTrue {
 						return true
 					}
 				}
@@ -262,7 +264,7 @@ var _ = Describe("VaultSecret controller", func() {
 				}
 
 				for _, condition := range rsCreated.Status.Conditions {
-					if condition.Type == "ReconcileSuccess" {
+					if condition.Type == vaultresourcecontroller.ReconcileSuccessful && condition.Status == metav1.ConditionTrue {
 						return true
 					}
 				}
@@ -290,7 +292,7 @@ var _ = Describe("VaultSecret controller", func() {
 				}
 
 				for _, condition := range created.Status.Conditions {
-					if condition.Type == "ReconcileSuccess" {
+					if condition.Type == vaultresourcecontroller.ReconcileSuccessful && condition.Status == metav1.ConditionTrue {
 						return true
 					}
 				}
